+++ src/angular-cesium.module.ts
@@ -30,6 +30,13 @@ import { AcPolygonDescComponent } from './components/ac-polygon-desc/ac-polygon-
 import { AcDefaultPlonterComponent } from './components/ac-default-plonter/ac-default-plonter.component';
 import { AcPolygonComponent } from './components/ac-polygon/ac-polygon.component';
 import { ViewersManagerService } from './services/viewers-service/viewers-manager.service';
+import { AcStaticEllipseDescComponent } from './components/static-dynamic/ac-static-ellipse-desc/ac-static-ellipse-desc.component';
+import { AcDynamicEllipseDescComponent } from './components/static-dynamic/ac-dynamic-ellipse-desc/ac-dynamic-ellipse-desc.component';
+import { AcDynamicPolylineDescComponent } from './components/static-dynamic/ac-dynamic-polyline-desc/ac-dynamic-polyline-desc.component';
+import { AcStaticPolygonDescComponent } from './components/static-dynamic/ac-static-polygon-desc/ac-static-polygon-desc.component';
+import { AcStaticCircleDescComponent } from './components/static-dynamic/ac-static-circle-desc/ac-static-circle-desc.component';
+import { AcDynamicCircleDescComponent } from './components/static-dynamic/ac-dynamic-circle-desc/ac-dynamic-circle-desc.component';
+import { AcStaticPolylineDescComponent } from './components/static-dynamic/ac-static-polyline-desc/ac-static-polyline-desc.component';
 
 @NgModule({
 	imports: [
@@ -62,6 +69,14 @@ import { ViewersManagerService } from './services/viewers-service/viewers-manage
     AcPolygonDescComponent,
     AcPolygonComponent,
     AcDefaultPlonterComponent,
+    
+    AcStaticEllipseDescComponent,
+  	AcDynamicEllipseDescComponent,
+  	AcDynamicPolylineDescComponent,
+		AcStaticPolylineDescComponent,
+		AcDynamicCircleDescComponent,
+    AcStaticCircleDescComponent,
+  	AcStaticPolygonDescComponent,
 	],
 	exports: [
 		AcMapComponent,
@@ -86,6 +101,14 @@ import { ViewersManagerService } from './services/viewers-service/viewers-manage
     AcPolygonDescComponent,
     AcPolygonComponent,
     AcDefaultPlonterComponent,
+    
+    AcStaticEllipseDescComponent,
+    AcDynamicEllipseDescComponent,
+    AcDynamicPolylineDescComponent,
+    AcStaticPolylineDescComponent,
+    AcDynamicCircleDescComponent,
+    AcStaticCircleDescComponent,
+    AcStaticPolygonDescComponent,
 	],
 	providers: [JsonMapper, CesiumProperties, GeoUtilsService, ViewerFactory, ViewersManagerService],
 })