+++ src/services/drawers/static-dynamic/static-primitive-drawer/static-primitive-drawer.service.ts
@@ -1,5 +1,5 @@
-import { PrimitivesDrawerService } from '../primitives-drawer/primitives-drawer.service';
-import { CesiumService } from '../../cesium/cesium.service';
+import { PrimitivesDrawerService } from '../../primitives-drawer/primitives-drawer.service';
+import { CesiumService } from '../../../cesium/cesium.service';
 
 /**
  * @deprecated