+++ CHANGELOG.md
@@ -9,4 +9,12 @@
 * Expose cesium viewer through ViewesManagerService and `ac-map`
 
 ## 0.0.14
+### Breaking changes 
+* deprecated: 
+  * AcStaticEllipseDescComponent, instead use `ac-ellipse-desc`
+  * AcDynamicEllipseDescComponent, instead use `ac-ellipse-desc`
+  * AcDynamicPolylineDescComponent, instead use `ac-polyline-desc`
+  * AcStaticPolylineDescComponent, instead use `ac-polyline-desc`
+  * AcStaticCircleDescComponent, instead use `ac-circle-desc`
+  * AcStaticPolygonDescComponent, instead use `ac-polygon-desc`
 * MapEventManagerService: EventResult.primitives changed to EventResult.cesiumEntities